packages/workers/requirements.txt
pika
python-dotenv
redis
hume
openai
pymongo
deepgram-sdk